Refactor watchlist and rollback right API call into WMFArticleDataController
