--- WMFData/Sources/WMFData/Models/Watchlist/WMFPageWatchStatus.swift
@@ -1,7 +1,0 @@
-import Foundation
-
-public struct WMFPageWatchStatus {
-    public let watched: Bool
-    public let watchlistExpiry: Date?
-    public let userHasRollbackRights: Bool?
-}
