+++ Wikipedia/Code/ArticleViewController+Watchlist.swift
@@ -2,36 +2,6 @@ import Foundation
 import WMFData
 
 extension ArticleViewController {
-    func loadWatchStatusAndUpdateToolbar() {
-        
-        guard let title = articleURL.wmf_title,
-        let siteURL = articleURL.wmf_site,
-        let project = WikimediaProject(siteURL: siteURL)?.wmfProject else {
-            return
-        }
-        
-        WMFWatchlistDataController().fetchWatchStatus(title: title, project: project) { result in
-            
-            DispatchQueue.main.async { [weak self] in
-                
-                guard let self else {
-                    return
-                }
-                
-                switch result {
-                case .success(let status):
-                    
-                    let needsWatchButton = !status.watched
-                    let needsUnwatchHalfButton = status.watched && status.watchlistExpiry != nil
-                    let needsUnwatchFullButton = status.watched && status.watchlistExpiry == nil
-                    
-                    self.toolbarController.updateMoreButton(needsWatchButton: needsWatchButton, needsUnwatchHalfButton: needsUnwatchHalfButton, needsUnwatchFullButton: needsUnwatchFullButton)
-                case .failure:
-                    break
-                }
-            }
-        }
-    }
     
     func watch() {
